+++ examples/5/main.py
@@ -11,6 +11,7 @@ from oamc.logging import enable_logging
 from oamc.post import Viewer
 
 DIR = Path(__file__).parent.resolve()
+CACHE_DIRECTORY = DIR / ".oamc" / "scene"
 
 
 # fmt: off
@@ -18,6 +19,9 @@ def main() -> None:
     enable_logging()
 
     print(BANNER)
+
+    if Viewer.view_cached(CACHE_DIRECTORY):
+        return
 
     parser = APDLParser(DIR / "c_beam_small_thick.dat")
 
@@ -124,6 +128,7 @@ def main() -> None:
         projection_method=ProjectionMethod.L2,
         opacity=0.3,
         paths=model.fibers_as_list,
+        cache_directory=CACHE_DIRECTORY,
     )